+++ DellOpenManage/Public/OME/Invoke-OMEReport.ps1
@@ -10,7 +10,7 @@
     $DeetsResp = Invoke-WebRequest -Uri $ReportDeets -UseBasicParsing -Headers $Headers -Method Get -ContentType $Type
     if ($DeetsResp.StatusCode -eq 200){
         $DeetsInfo = $DeetsResp.Content | ConvertFrom-Json
-        $ColumnNames = $DeetsInfo.ColumnNames.Name
+        $ColumnNames = $DeetsInfo.ColumnNames | Sort-Object Sequence | ForEach-Object{$_.Name}
         Write-Verbose "Extracting results for report ($($ReportId))"
         $ResultUrl = $BaseUri + "/api/ReportService/ReportDefs($($ReportId))/ReportResults/ResultRows"